tests/wpt/tests/css/css-images/animation/image-slice-interpolation-math-functions-tentative.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>border-image-slice interpolation with css math functions</title>
    <link rel="author" title="Daniil Sakhapov" href="mailto:sakhapov@chromium.org">
    <link rel="help" href="https://drafts.csswg.org/css-backgrounds/#the-border-image-slice">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>
  <style>
    :root {
      font-size: 10px;
    }
  </style>
  <body>
    <script>
      test_interpolation({
        property: 'border-image-slice',
        from: '100',
        to: 'calc(sign(20rem - 20px) * 180)',
      }, [
        {at: -1, expect: '20'},
        {at: 0, expect: '100'},
        {at: 0.125, expect: '110'},
        {at: 0.875, expect: '170'},
        {at: 1, expect: '180'},
        {at: 2, expect: '260'}
      ]);

      test_interpolation({
        property: 'border-image-slice',
        from: '100%',
        to: 'calc(sign(20rem - 20px) * 180%)',
      }, [
        {at: -1, expect: '20%'},
        {at: 0, expect: '100%'},
        {at: 0.125, expect: '110%'},
        {at: 0.875, expect: '170%'},
        {at: 1, expect: '180%'},
        {at: 2, expect: '260%'}
      ]);

      test_interpolation({
        property: 'border-image-slice',
        from: 'calc(sign(20rem - 20px) * 100%)',
        to: 'calc(progress(10rem from 20px to 100px) * 180%)',
      }, [
        {at: -1, expect: '20%'},
        {at: 0, expect: '100%'},
        {at: 0.125, expect: '110%'},
        {at: 0.875, expect: '170%'},
        {at: 1, expect: '180%'},
        {at: 2, expect: '260%'}
      ]);
    </script>
  </body>
</html>
